+++ packages/app/screens/LibraryScreen/DownloadTab.tsx
@@ -1,21 +1,27 @@
 import { View, Text, FlatList } from "native-base";
 import React from "react";
 
-import { HistoryComicT, historySelector } from "app/store/historySlice";
+import {
+  HistoryComicT,
+  historySelector,
+  selectDownloadedComics
+} from "app/store/historySlice";
 import { useAppDispatch, useAppSelector } from "app/store/hooks";
 import { NextLink } from "app/components/NextLink";
 import LibraryList from "./LibraryList";
 import { navigate } from "app/navigators";
 import { LibraryContext } from "./LibraryContext";
-import { Download } from "app/components/EmptyPage";
+import { Download, Loading } from "app/components/EmptyPage";
+import { useIsFocused } from "../../hooks/useIsFocused";
 
 interface Props {}
 
 export const DownloadTab = (props: Props) => {
   const dispatch = useAppDispatch();
-  const { comics, downloadComics, downloadCpt } =
-    useAppSelector(historySelector);
-
+  const { isFocused } = useIsFocused();
+  const data = useAppSelector((state) =>
+    selectDownloadedComics(state, isFocused)
+  );
   const { showModal } = React.useContext(LibraryContext);
 
   return (
@@ -36,30 +42,35 @@ export const DownloadTab = (props: Props) => {
         keyExtractor={(item) => item.title}
       /> */}
 
-      <LibraryList
-        customLoadingComponent={Download}
-        data={
-          (downloadComics?.map((str) => comics[str]) as HistoryComicT[]) || []
-        }
-        addonFieldName={"Downloaded:"}
-        addonFieldExtractor={(comic) =>
-          comic.chapters
-            .map((cpt) => downloadCpt[cpt.path])
-            .filter((p) => p)
-            .length.toString()
-        }
-        onPress={(comic) =>
-          navigate("offline-comic-screen", {
-            path: comic.path || ""
-          })
-        }
-        onLongPress={(comic) => {
-          showModal &&
-            showModal(true, comic.path, () => (path) => {
-              console.log("test remove downloaded chapter " + path);
-            });
-        }}
-      />
+      {isFocused ? (
+        <LibraryList
+          customLoadingComponent={Download}
+          data={data}
+          addonFieldName={"Downloaded:"}
+          addonFieldExtractor={
+            (comic) =>
+              // FIXME: CHANGE EXTRACTOR TO FIT WITH NEW STRUCTURE
+              "Fixme"
+            // comic.chapters
+            //   .map((cpt) => downloadCpt[cpt.path])
+            //   .filter((p) => p)
+            //   .length.toString()
+          }
+          onPress={(comic) =>
+            navigate("offline-comic-screen", {
+              path: comic.path || ""
+            })
+          }
+          onLongPress={(comic) => {
+            showModal &&
+              showModal(true, comic.path, () => (path) => {
+                console.log("test remove downloaded chapter " + path);
+              });
+          }}
+        />
+      ) : (
+        <Loading />
+      )}
     </View>
   );
 };