+++ packages/app/screens/LibraryScreen/RecentTab.tsx
@@ -13,11 +13,21 @@ import { View } from "native-base";
 import LibraryList from "./LibraryList";
 import ConfirmModal from "app/components/ConfirmModal";
 import { LibraryContext } from "./LibraryContext";
+import { useIsFocused } from "../../hooks/useIsFocused";
+import { Loading } from "../../components/EmptyPage";
 
 export const RecentTab: React.FunctionComponent<RecentTabProps> = (props) => {
   // const history = useAppSelector(historySelector);
-  const data = useAppSelector(selectReadComics);
+
+  const { isFocused } = useIsFocused();
+  console.log(isFocused);
+  const data = useAppSelector((state) => selectReadComics(state, isFocused));
   const dispatch = useAppDispatch();
+
+  // const renderData = React.useMemo(() => {
+  //   if(isFocused === true) return data;
+  //   else return [];
+  // }, [isFocused]);
 
   const { showModal } = React.useContext(LibraryContext);
 
@@ -43,12 +53,16 @@ export const RecentTab: React.FunctionComponent<RecentTabProps> = (props) => {
 
   return (
     <View style={{ flex: 1 }} bg={"warmGray.50"} _dark={{ bg: "warmGray.900" }}>
-      <LibraryList
-        data={data}
-        addonFieldName={"Lasted read: "}
-        addonFieldExtractor={fieldExtractor}
-        onLongPress={onLongPress}
-      />
+      {isFocused ? (
+        <LibraryList
+          data={data}
+          addonFieldName={"Lasted read: "}
+          addonFieldExtractor={fieldExtractor}
+          onLongPress={onLongPress}
+        />
+      ) : (
+        <Loading />
+      )}
     </View>
   );
 };